+++ src/clj/rems/db/form.clj
@@ -114,8 +114,7 @@
 
 (defn create-form! [user-id form]
   (let [organization (:form/organization form)]
-    (or (util/forbidden-organization? user-id organization)
-        (validation-error form)
+    (or (util/forbidden-organization-error organization)
         (let [form-id (:id (db/save-form-template! {:organization organization
                                                     :title (:form/title form)
                                                     :user user-id
@@ -126,7 +125,7 @@
 (defn edit-form! [user-id form]
   (let [form-id (:form/id form)
         organization (:form/organization form)]
-    (or (util/forbidden-organization? user-id organization)
+    (or (util/forbidden-organization-error organization)
         (form-in-use-error form-id)
         (do (db/edit-form-template! {:id form-id
                                      :organization organization