+++ test/clj/rems/api/services/test_catalogue.clj
@@ -7,7 +7,8 @@
             [rems.api.services.resource :as resource]
             [rems.api.services.workflow :as workflow]
             [rems.db.test-data :as test-data]
-            [rems.db.testing :refer [caches-fixture rollback-db-fixture test-db-fixture]]))
+            [rems.db.testing :refer [caches-fixture rollback-db-fixture test-db-fixture]]
+            [rems.testing-util :refer [with-user]]))
 
 (use-fixtures :once test-db-fixture caches-fixture)
 (use-fixtures :each rollback-db-fixture)
@@ -17,7 +18,8 @@
       (select-keys [:enabled :archived])))
 
 (deftest catalogue-item-enabled-archived-test
-  (let [form-id (test-data/create-form! {})
+  (let [owner "owner"
+        form-id (test-data/create-form! {})
         lic-id (test-data/create-license! {})
         res-id (test-data/create-resource! {:license-ids [lic-id]})
         workflow-id (test-data/create-workflow! {})
@@ -26,16 +28,16 @@
                                                    :workflow-id workflow-id})
         item-id2 (test-data/create-catalogue-item! {})
 
-        enable-catalogue-item!
-        #(catalogue/set-catalogue-item-enabled! {:id item-id
-                                                 :enabled %})
-        archive-catalogue-item!
-        #(catalogue/set-catalogue-item-archived! {:id item-id
-                                                  :archived %})
+        enable-catalogue-item! #(catalogue/set-catalogue-item-enabled! {:id item-id
+                                                                        :enabled %})
+        archive-catalogue-item! #(with-user owner
+                                   (catalogue/set-catalogue-item-archived! {:id item-id
+                                                                            :archived %}))
         archive-form! #(form/set-form-archived! {:id form-id
                                                  :archived %})
-        archive-license! #(licenses/set-license-archived! {:id lic-id
-                                                           :archived %})
+        archive-license! #(with-user owner
+                            (licenses/set-license-archived! {:id lic-id
+                                                             :archived %}))
         archive-resource! #(resource/set-resource-archived! {:id res-id
                                                              :archived %})
         archive-workflow! #(workflow/set-workflow-archived! {:id workflow-id
@@ -77,7 +79,8 @@
       (enable-catalogue-item! true)
       (archive-catalogue-item! true)
       (catalogue/set-catalogue-item-enabled! {:id item-id2 :enabled false})
-      (catalogue/set-catalogue-item-archived! {:id item-id2 :archived false})
+      (with-user owner
+        (catalogue/set-catalogue-item-archived! {:id item-id2 :archived false}))
       (is (= {:enabled true
               :archived true}
              (status-flags item-id)))
@@ -135,14 +138,16 @@
     (is (= "Uusi nimi" (get-in new-item [:localizations :fi :title])))))
 
 (deftest test-get-localized-catalogue-items
-  (let [item-id (test-data/create-catalogue-item! {})]
+  (let [owner "owner"
+        item-id (test-data/create-catalogue-item! {})]
 
     (testing "find all"
       (is (= [item-id] (map :id (catalogue/get-localized-catalogue-items)))))
 
     (testing "archived catalogue items"
-      (catalogue/set-catalogue-item-archived! {:id item-id
-                                               :archived true})
+      (with-user owner
+        (catalogue/set-catalogue-item-archived! {:id item-id
+                                                 :archived true}))
       (is (= [] (map :id (catalogue/get-localized-catalogue-items))))
       (is (= [item-id] (map :id (catalogue/get-localized-catalogue-items {:archived true}))))
       (is (= [] (map :id (catalogue/get-localized-catalogue-items {:archived false})))))))