+++ test/clj/rems/api/test_entitlements.clj
@@ -48,12 +48,14 @@
                                            :end nil
                                            :mail "malice@example.com"}
                                           (dissoc x :start :application-id)))
-                                   (is (valid-date? (:start x))))]
+                                   (is (valid-date? (:start x))))
+        sort-entitlements (partial sort-by (juxt :mail :end))]
     (testing "all"
       (let [data (-> (request :get "/api/entitlements")
                      (authenticate api-key "developer")
                      handler
-                     read-ok-body)]
+                     read-ok-body
+                     sort-entitlements)]
         (is (= 2 (count data)))
         (check-alice-entitlement (first data))
         (check-malice-entitlement (second data))))
@@ -63,7 +65,8 @@
         (let [data (-> (request :get "/api/entitlements")
                        (authenticate api-key userid)
                        handler
-                       read-ok-body)]
+                       read-ok-body
+                       sort-entitlements)]
           (is (= 2 (count data)))
           (check-alice-entitlement (first data))
           (check-malice-entitlement (second data)))))
@@ -72,7 +75,8 @@
       (let [data (-> (request :get "/api/entitlements?resource=urn:nbn:fi:lb-201403262")
                      (authenticate api-key "developer")
                      handler
-                     read-ok-body)]
+                     read-ok-body
+                     sort-entitlements)]
         (is (= 2 (count data)))
         (check-alice-entitlement (first data))
         (check-malice-entitlement (second data))))
@@ -97,7 +101,8 @@
       (let [data (-> (request :get "/api/entitlements?expired=true")
                      (authenticate api-key "owner")
                      handler
-                     read-ok-body)]
+                     read-ok-body
+                     sort-entitlements)]
         (is (= 3 (count data)))
         (check-alice-entitlement (first data))
         (check-alice-expired-entitlement (second data))