+++ test/clj/rems/test_browser.clj
@@ -39,9 +39,10 @@
 
 (defn fixture-standalone [f]
   (mount/start)
-  (migrations/migrate ["reset"] (select-keys rems.config/env [:database-url]))
+  (migrations/migrate ["migrate"] (select-keys rems.config/env [:database-url]))
   (test-data/create-test-data!)
   (f)
+  (migrations/migrate ["reset"] (select-keys rems.config/env [:database-url]))
   (mount/stop))
 
 (defn smoke-test [f]